model_training/pyproject.toml
[tool.poetry]
name = "model"
version = "0.1.0"
description = "model training env settings"
authors = ["chaea <chaea11s0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torchvision = "^0.17.1"
pandas = "^2.2.1"
mlflow = "^2.11.1"
tqdm = "^4.66.2"
yfinance = "^0.2.37"
glances = {extras = ["all"], version = "^3.4.0.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
